test primary tools and fix issues
